# Improved performance of `String.repeat()`