@Mickeon 2022-08-16
Would it be a reasonable loss of performance to switch between the two algorithms depending on how long the String is?