@MewPurPur 2022-08-16
Sure, could do it after the basic case checks. The optimization performs better for longer strings as well as for more repeats, so I'd have to draw the line for each individual string length, where length is 1 or more, and repeats are 2 or more.