@kleonc 2022-08-16
`String::repeat` seems to be kinda the same as `Image::fill` (string = color, repetition count = image size), I wonder how [approach taken there](https://github.com/godotengine/godot/pull/52456/commits/92ca300ab17faa044b92d753a05a71f892d7bede) would perform here. If you want you could benchmark it, I think something like this would be equivalent (haven't tested it though):
```cpp
String String::repeat(int p_count) const {
	ERR_FAIL_COND_V_MSG(p_count < 0, "", "Parameter count should be a positive number.");

	if (p_count == 0 || is_empty()) {
		return "";
	}

	if (p_count == 1) {
		return *this;
	}	
	
	String new_string = *this;
	int len = length();
	int len_bytes = len * sizeof(char32_t);
	new_string.resize(len * p_count + 1);
	new_string[len * p_count] = 0;	
	
	char32_t *data_ptr = new_string->ptrw();	
	
	int offset = 1;
	for (int stride = 1; offset + stride <= p_count; stride *= 2) {
		memcpy(data_ptr + offset * len_bytes, data_ptr, stride * len_bytes);
		offset += stride;
	}
	if (offset < p_count) {
		memcpy(data_ptr + offset * len_bytes, data_ptr, (p_count - offset) * len_bytes);
	}
	
	return new_string;
}
```