@kleonc 2022-08-16
> @kleonc Couldn't get it to work, might just be my shortcomings with C++.

I've edited it, should be fine now (I mean engine is not crashing at startup).