@kleonc 2022-08-16
@MewPurPur You need to fix the formatting (you have spaces instead of tabs), see [Code style guidelines](https://docs.godotengine.org/en/latest/community/contributing/code_style_guidelines.html#code-style-guidelines) for how-to (or go directly to [hooks](https://github.com/godotengine/godot/tree/master/misc/hooks)).