@MewPurPur 2022-08-17
No idea why the checks don't pass. But I'm still want to first re-implement this with memcpy. I've figured the constant resizes from the concatenation are most likely the reason for the performance drop on small strings.