@kleonc 2022-08-17
> No idea why the checks don't pass.

It happens because `String::operator+=(const String &)` doesn't properly handle case when the same string is passed in (e.g. like in `str += str`).  Basically in this line:
https://github.com/godotengine/godot/blob/dbd15243621ec595742b18abc4c26f3cb2e00f3d/core/string/ustring.cpp#L537
in such case destination and source memory blocks overlap as `src + rhs_len * sizeof(char32_t) == dst` (last `char32_t` to copy from `src` is at the same address as first `char32_t` in `dst`). And that's an undefined behavior as `memcpy` requires these to not overlap.