@MewPurPur 2022-08-18
Ended up doing a condition for different lengths after all. Thankfully, it turns out for strings longer than 4 characters, this solution is always beneficial.