@kleonc 2022-08-19
> Ended up doing a condition for different lengths after all. Thankfully, it turns out for strings longer than 4 characters, this solution is always beneficial.

Looks way overcomplicated to me. I've edited [my previous comment](https://github.com/godotengine/godot/pull/64489#issuecomment-1216706064) with a new version using `memcpy`, check it out. This one should be correct. :wink: