@MewPurPur 2022-08-19
@kleonc Oh, my, god.

The performance difference is stunning, especially compared to how it used to be. All of the tests now take between 5-8 ms in the benchmark.