@akien-mga 2022-08-22
> I'm not sure why the godot-cpp test is failing. I restarted the build, if it fails again it might require a rebase if somehow this PR was made at a time where godot-cpp was in a broken state.

This will need a rebase indeed to fix the CI.