@MewPurPur 2022-08-22
Never done rebases, hope I did well. Also I removed the conditions for basic cases since the default algorithm handles them fast now.