@Calinou 2022-08-23
> Benchmark (updated):

@MewPurPur Do you have the benchmark script available? I'm working on backporting this optimization to `3.x` :slightly_smiling_face: