@MewPurPur 2022-08-23
@Calinou Yes, just gimme a few mins to get to my laptop!